// src/art/enemyFrameGeometry.generated.ts
export const enemyFrameGeometry = {
  "cave-brute": {
    "idle": [
      624,
      724
    ],
    "run": [
      720,
      724
    ],
    "anticipation": [
      656,
      724
    ],
    "attack": [
      672,
      724
    ],
    "attack-fx": [
      832,
      724
    ],
    "dodge": [
      848,
      724
    ],
    "block": [
      608,
      724
    ],
    "hit": [
      784,
      724
    ],
    "ko": [
      832,
      724
    ]
  },
  "tribal-warrior": {
    "idle": [
      592,
      724
    ],
    "run": [
      768,
      724
    ],
    "anticipation": [
      880,
      724
    ],
    "attack": [
      784,
      724
    ],
    "attack-fx": [
      704,
      724
    ],
    "dodge": [
      768,
      724
    ],
    "block": [
      672,
      724
    ],
    "hit": [
      672,
      724
    ],
    "ko": [
      736,
      724
    ]
  },
  "tribal-hunter": {
    "idle": [
      592,
      724
    ],
    "run": [
      768,
      724
    ],
    "anticipation": [
      880,
      724
    ],
    "attack": [
      784,
      724
    ],
    "attack-fx": [
      704,
      724
    ],
    "dodge": [
      768,
      724
    ],
    "block": [
      672,
      724
    ],
    "hit": [
      672,
      724
    ],
    "ko": [
      736,
      724
    ]
  },
  "raptor": {
    "idle": [
      640,
      724
    ],
    "run": [
      704,
      724
    ],
    "anticipation": [
      800,
      724
    ],
    "attack": [
      720,
      724
    ],
    "attack-fx": [
      656,
      724
    ],
    "dodge": [
      704,
      724
    ],
    "block": [
      656,
      724
    ],
    "hit": [
      688,
      724
    ],
    "ko": [
      704,
      724
    ]
  },
  "shaman": {
    "idle": [
      608,
      724
    ],
    "run": [
      656,
      724
    ],
    "anticipation": [
      672,
      724
    ],
    "attack": [
      880,
      724
    ],
    "attack-fx": [
      656,
      724
    ],
    "dodge": [
      720,
      724
    ],
    "block": [
      704,
      724
    ],
    "hit": [
      800,
      724
    ],
    "ko": [
      608,
      724
    ]
  },
  "smilodon": {
    "idle": [
      627,
      627
    ],
    "run": [
      736,
      627
    ],
    "anticipation": [
      627,
      627
    ],
    "attack": [
      752,
      627
    ],
    "attack-fx": [
      784,
      627
    ],
    "dodge": [
      912,
      627
    ],
    "block": [
      627,
      627
    ],
    "hit": [
      784,
      627
    ],
    "ko": [
      800,
      627
    ]
  },
  "mammoth": {
    "idle": [
      672,
      724
    ],
    "run": [
      656,
      724
    ],
    "anticipation": [
      672,
      724
    ],
    "attack": [
      752,
      724
    ],
    "attack-fx": [
      784,
      724
    ],
    "dodge": [
      640,
      724
    ],
    "block": [
      640,
      724
    ],
    "hit": [
      688,
      724
    ],
    "ko": [
      720,
      724
    ]
  }
} as const
export const enemyFramePaintedSize = {
  "cave-brute": {
    "idle": [
      552,
      594
    ],
    "run": [
      610,
      568
    ],
    "anticipation": [
      576,
      611
    ],
    "attack": [
      565,
      692
    ],
    "attack-fx": [
      687,
      550
    ],
    "dodge": [
      673,
      561
    ],
    "block": [
      532,
      679
    ],
    "hit": [
      613,
      646
    ],
    "ko": [
      648,
      470
    ]
  },
  "tribal-warrior": {
    "idle": [
      494,
      626
    ],
    "run": [
      630,
      553
    ],
    "anticipation": [
      772,
      525
    ],
    "attack": [
      665,
      594
    ],
    "attack-fx": [
      617,
      591
    ],
    "dodge": [
      614,
      507
    ],
    "block": [
      581,
      651
    ],
    "hit": [
      535,
      615
    ],
    "ko": [
      602,
      556
    ]
  },
  "tribal-hunter": {
    "idle": [
      494,
      626
    ],
    "run": [
      630,
      553
    ],
    "anticipation": [
      772,
      525
    ],
    "attack": [
      665,
      594
    ],
    "attack-fx": [
      617,
      591
    ],
    "dodge": [
      614,
      507
    ],
    "block": [
      581,
      651
    ],
    "hit": [
      535,
      615
    ],
    "ko": [
      602,
      556
    ]
  },
  "raptor": {
    "idle": [
      542,
      581
    ],
    "run": [
      582,
      428
    ],
    "anticipation": [
      658,
      483
    ],
    "attack": [
      592,
      532
    ],
    "attack-fx": [
      547,
      480
    ],
    "dodge": [
      587,
      500
    ],
    "block": [
      550,
      565
    ],
    "hit": [
      588,
      575
    ],
    "ko": [
      569,
      573
    ]
  },
  "shaman": {
    "idle": [
      554,
      681
    ],
    "run": [
      544,
      648
    ],
    "anticipation": [
      602,
      703
    ],
    "attack": [
      742,
      588
    ],
    "attack-fx": [
      527,
      599
    ],
    "dodge": [
      579,
      655
    ],
    "block": [
      604,
      601
    ],
    "hit": [
      622,
      687
    ],
    "ko": [
      543,
      514
    ]
  },
  "smilodon": {
    "idle": [
      598,
      444
    ],
    "run": [
      671,
      356
    ],
    "anticipation": [
      590,
      401
    ],
    "attack": [
      662,
      373
    ],
    "attack-fx": [
      720,
      475
    ],
    "dodge": [
      685,
      376
    ],
    "block": [
      620,
      409
    ],
    "hit": [
      641,
      517
    ],
    "ko": [
      614,
      482
    ]
  },
  "mammoth": {
    "idle": [
      578,
      545
    ],
    "run": [
      588,
      471
    ],
    "anticipation": [
      562,
      486
    ],
    "attack": [
      613,
      589
    ],
    "attack-fx": [
      605,
      555
    ],
    "dodge": [
      557,
      504
    ],
    "block": [
      551,
      483
    ],
    "hit": [
      574,
      546
    ],
    "ko": [
      590,
      454
    ]
  }
} as const